Separated ViewController and ViewModel in ViewController file.
